test(engine): cover optional-cost AI contract integration
